Permissionクラスを削除
クラスとしての役割が少ないため削除しました。
